Calculate demand quantities for a single set of inputs

This is a helper function for the mapply calls in food.dmnd

calc1q(Ps, Pn, Y, eps, Ysterm, Ynterm, Acoef, psscl, pnscl)

Arguments

Ps
Price of staple foods

Pn
Price of nonstaple foods

Y
Per-capita income

eps
Matrix of Marshall elasticities

Ysterm
Income term in the demand equation for staples

Ynterm
Income term in the demand equation for nonstaples

Acoef
Leading multiplier parameter.

Value

Quantities of staples and non-staples (Qs and Qn)