Create dataset with observational error for log likelihood calculation using clustering.

Create dataset with observational error for log likelihood calculation using clustering.

create_dataset_for_parameter_fit(
  min_price_pd = 20,
  min_cal_fd = 1000,
  min_clusters = 300,
  lower_limit_sigma = 0.01,
  data = NULL,
  outdir = "tests/testthat/test_outputs/",
  print_progress = FALSE
)

Arguments

min_cal_fd
Minimum calories for the food demand model

min_clusters
Minimum number of clusters to be generated by clustering algoritm. It is recommended to not lowewr this parameter below 20.

lower_limit_sigma
Lower limit for sigma values calculated

data
A data.frame or data.table with the raw data. Data should contain following names,

s_cal_pcap_day_thous (Containing 1000 calories per capita per day for staples)

ns_cal_pcap_day_thous (Containing 1000 calories per capita per day for non-staples)

gdp_pcap_thous (Containing GDP per capita in thous USD)

s_usd_p1000cal (Price of 1000 calories for staples per person per day)

ns_usd_p1000cal (Price of 1000 calories for non-staples per person per day)

outdir
Directory to store output csv. Default is set to test_output folder.

print_progress
A parameter that allows the user to track progress of function.

min_price_fd
Highest possible price of staples and non-staples

original_param_vector
Original parameter vector to be used as the starting point for the optimization.These parameters are taken from Edmonds et al 2017

optim_method
The optimization method to be used for maximization of the log likelihood. The default is set to BFGS

Value

A dataframe called Processed_Data_for_MC with the following columns

sig2Qn- which is the observational error for non-staples

sig2Qs- which is the observational error for staples

Author

KBN 2020